requirements.txt
anyio==3.6.2
asttokens==2.2.1
attrs==22.2.0
blinker==1.5
CacheControl==0.12.11
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
decorator==5.1.1
distlib==0.3.6
dnspython==2.2.1
executing==1.2.0
fastapi==0.88.0
filelock==3.9.0
findpython==0.2.2
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
icecream==2.1.3
idna==3.4
iniconfig==1.1.1
installer==0.6.0
lockfile==0.12.2
msgpack==1.0.4
packaging==21.3
pdm==2.3.4
platformdirs==2.6.2
pluggy==1.0.0
pydantic==1.10.2
Pygments==2.13.0
pymongo==4.3.3
pyparsing==3.0.9
pyproject_hooks==1.0.0
pytest==7.2.0
pytest-asyncio==0.20.3
python-dotenv==0.21.0
requests==2.28.1
requests-toolbelt==0.10.1
resolvelib==0.9.0
rfc3986==1.5.0
rich==12.6.0
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomlkit==0.11.6
typing_extensions==4.4.0
unearth==0.7.0
urllib3==1.26.13
uvicorn==0.20.0
virtualenv==20.17.1
